requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
pyTelegramBotAPI==4.16.1
python-decouple==3.8
requests==2.31.0
soupsieve==2.5
urllib3==2.2.1
